// src/room/interfaces.ts
export interface RoomConfig {
  // TODO: remove threshold configs if they are not necessary
  SPAWN_WITHDRAW_THRESHOLD: number;
  CONTAINER_WITHDRAW_THRESHOLD: number;
  STORAGE_WITHDRAW_THRESHOLD: number;
  tower: any;
  spawn: any;
  harvester: any;
  miner: any;
  extractor: any;
  upgrader: any;
  builder: any;
  repairer: any;
  transferrer: any;
  army: any;
}